pkg/ansible/proxy: move pkg/ansible/proxy to pkg/proxy
